fix: restructure tests and fix cluster error
