+++ package.json
@@ -9,10 +9,9 @@
   "scripts": {
     "lint": "eslint",
     "lint:fix": "eslint --fix",
-    "unit": "node --test test.js",
+    "unit": "node --test test/*.js",
     "test": "npm run lint && npm run unit",
-    "test:clusters": "node --test test-clusters.js",
-    "coverage": "nyc --reporter=lcov node --test test.js",
+    "coverage": "c8 --reporter=lcov npm run unit",
     "license-checker": "license-checker --production --onlyAllow='MIT;ISC;BSD-3-Clause;BSD-2-Clause;Apache-2.0;Apache*'",
     "release": "read -p 'GITHUB_TOKEN: ' GITHUB_TOKEN && export GITHUB_TOKEN=$GITHUB_TOKEN && release-it --disable-metrics",
     "redis": "cd docker;docker compose up"
@@ -54,19 +53,19 @@
   "homepage": "https://github.com/moscajs/aedes-persistence-redis#readme",
   "devDependencies": {
     "@fastify/pre-commit": "^2.2.0",
-    "eslint": "^9.21.0",
+    "c8": "^10.1.3",
+    "eslint": "^9.25.1",
     "fastq": "^1.19.1",
     "license-checker": "^25.0.1",
-    "mqemitter-redis": "^6.1.0",
-    "mqtt": "^5.10.4",
+    "mqemitter-redis": "^7.0.0",
+    "mqtt": "^5.11.1",
     "neostandard": "^0.12.1",
-    "nyc": "^17.1.0",
-    "release-it": "^18.1.2"
+    "release-it": "^19.0.1"
   },
   "dependencies": {
     "aedes-cached-persistence": "^10.0.0",
     "hashlru": "^2.3.0",
-    "ioredis": "^5.5.0",
+    "ioredis": "^5.6.1",
     "msgpack-lite": "^0.1.26",
     "qlobber": "^8.0.1"
   }
